Refactor runvm as a standalone command
New command names:
 - `trace record` -> `aida-trace record`
 - `trace replay` -> `aida-trace replay`
 - `trace run-vm` -> `aida-runvm`
