+++ Makefile
@@ -20,7 +20,7 @@ GOPROXY ?= "https://proxy.golang.org,direct"
 
 .PHONY: all clean help test
 
-all: gen-world-state trace aida-stochastic
+all: gen-world-state aida-trace aida-runvm aida-stochastic
 
 gen-world-state:
 	@go build \
@@ -39,15 +39,25 @@ aida-stochastic:
        	-o $(GO_BIN)/aida-stochastic \
 	./cmd/stochastic-cli
 
-trace:
+aida-trace:
 	@cd carmen/go/lib ; \
 	./build_libcarmen.sh ; \
 	cd ../../.. ; \
 	GOPROXY=$(GOPROXY) \
 	GOPRIVATE=github.com/Fantom-foundation/Carmen,github.com/Fantom-foundation/go-opera-fvm \
 	go build -ldflags "-s -w -X 'github.com/Fantom-foundation/Aida/utils.GitCommit=$(BUILD_COMMIT)'" \
-       	-o $(GO_BIN)/trace \
+	-o $(GO_BIN)/aida-trace \
 	./cmd/trace-cli
+
+aida-runvm:
+	@cd carmen/go/lib ; \
+	./build_libcarmen.sh ; \
+	cd ../../.. ; \
+	GOPROXY=$(GOPROXY) \
+	GOPRIVATE=github.com/Fantom-foundation/Carmen,github.com/Fantom-foundation/go-opera-fvm \
+	go build -ldflags "-s -w -X 'github.com/Fantom-foundation/Aida/utils.GitCommit=$(BUILD_COMMIT)'" \
+	-o $(GO_BIN)/aida-runvm \
+	./cmd/runvm-cli
 
 test:
 	@go test ./...
