+++ README.md
@@ -65,7 +65,7 @@ Trace cli tool provides storage trace recording and replaying functionality.
 ### Trace Record
 **Run**
 
-`./build/trace record 5000000 5100000`
+`./build/aida-trace record 5000000 5100000`
 simulates transaction execution from block 5,000,000 to (and including) block 5,100,000 using [substate](github.com/Fantom-foundation/substate-cli). Storage operations executed during transaction processing are recorded into a compressed file format.
 
 **Options**
@@ -81,7 +81,7 @@ simulates transaction execution from block 5,000,000 to (and including) block 5,
 
 **Run**
 
-`./build/trace replay --worldstatedir path/to/world-state 5050000 5100000`
+`./build/aida-trace replay --worldstatedir path/to/world-state 5050000 5100000`
 reads the recorded traces and re-executes state operations from block 5,050,000 to 5,100,000. The tool initializes stateDB with accounts in the world state from option `--worldstatedir`. The storage operations are executed and update the stateDB sequentially in the order they were recorded.
 
 **Options**
@@ -108,7 +108,7 @@ reads the recorded traces and re-executes state operations from block 5,050,000 
 
 **Run**
 
-`./build/trace replay-substate 5050000 5100000`
+`./build/aida-trace replay-substate 5050000 5100000`
 reads the recorded traces and re-executes state operations from block 5,050,000 to 5,100,000. The storage operations are executed sequentially in the order they were recorded. The tool iterates through substates to construct a partial stateDB such that the replayed storage operations can simulate read/write with actual data.
 
 **Options**
@@ -136,7 +136,7 @@ reads the recorded traces and re-executes state operations from block 5,050,000 
 
 **Run**
 
-`./build/trace run-vm --updatedir path/to/updatedb --db-impl [geth/carmen/memory/flat] 4564026 5000000`
+`./build/aida-runvm --updatedir path/to/updatedb --db-impl [geth/carmen/memory/flat] 4564026 5000000`
 executes transactions from block 4,564,026 to 5,000,000. The tool initializes stateDB with accounts in the world state from option `--worldstatedir`. Each transaction calls VM which issues a series of StateDB operations to a selected storage system.
 
 **Options**
@@ -169,7 +169,7 @@ executes transactions from block 4,564,026 to 5,000,000. The tool initializes st
 
 **Run**
 
-`./build/trace gen-update-set --worldstatedir path/to/world-state --updatedir path/to/updatedb 4564026 41000000 1000000`
+`./build/aida-trace gen-update-set --worldstatedir path/to/world-state --updatedir path/to/updatedb 4564026 41000000 1000000`
 generates piecewise update-sets (merges of output substates) at every 1000000 blocks starting from block 4564026 to block 41000000 and stores them in updateDB. SubstateAlloc of block 4564025 from the world state is reocrded as the first update-set if --worldstatedir is provided. The subsequence update-sets happen at block 5000000 and every 1000000 blocks afterwards. 
 
 **Options**
