+++ cmd/runvm-cli/runvm/proxy_profiler.go
@@ -1,4 +1,4 @@
-package tracer
+package runvm
 
 import (
 	"math/big"
