+++ cmd/trace-cli/trace/proxy_recorder.go
@@ -1,4 +1,4 @@
-package tracer
+package trace
 
 import (
 	"math/big"
